# Don't record measurements with no subscription