--- a/stats/view/worker_commands.go
+++ b/stats/view/worker_commands.go
@@ -22,6 +22,7 @@ import (
 	"time"
 
 	"go.opencensus.io/stats"
+	"go.opencensus.io/stats/internal"
 	"go.opencensus.io/tag"
 )
 
@@ -57,6 +58,7 @@ func (cmd *subscribeToViewReq) handleCommand(w *worker) {
 			errstr = append(errstr, fmt.Sprintf("%s: %v", view.Name, err))
 			continue
 		}
+		internal.SubscriptionReporter(view.Measure.Name())
 		vi.subscribe()
 	}
 	if len(errstr) > 0 {
@@ -135,6 +137,9 @@ type recordReq struct {
 
 func (cmd *recordReq) handleCommand(w *worker) {
 	for _, m := range cmd.ms {
+		if (m == stats.Measurement{}) { // not subscribed
+			continue
+		}
 		ref := w.getMeasureRef(m.Measure().Name())
 		for v := range ref.views {
 			v.addSample(cmd.tm, m.Value())