Suggested by Ramon, immediately dropping the measurements from
the measures with no subscription rather than doing it at a later time
significantly improves the performance on the critical path.

Before:
BenchmarkRecord0-8            	500000000	         3.09 ns/op
BenchmarkRecord1-8            	 5000000	       366 ns/op
BenchmarkRecord8-8            	 3000000	       412 ns/op
BenchmarkRecord8_Parallel-8   	 2000000	       804 ns/op
BenchmarkRecord8_8Tags-8      	 3000000	       415 ns/op

After:
BenchmarkRecord0-8            	500000000	         3.75 ns/op
BenchmarkRecord1-8            	30000000	        39.7 ns/op
BenchmarkRecord8-8            	20000000	       101 ns/op
BenchmarkRecord8_Parallel-8   	30000000	        40.1 ns/op
BenchmarkRecord8_8Tags-8      	20000000	       102 ns/op