@rakyll 2018-03-16
We can actually easily use the ref count.